Minor cleanup: count contribution per texel in bits, not bytes.

+++ lib/VolumeViz/render/2D/Cvr2DTexSubPage.h
@@ -36,7 +36,7 @@ private:
   static unsigned int nroftexels;
   static unsigned int texmembytes;
   static SbBool detectedtextureswapping;
-  float texmultfactor;
+  unsigned int texmultfactor;
 };
